requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
openai>=1.3.0
python-dotenv>=1.0.0
pydantic>=2.5.0
schedule>=1.2.0
sqlalchemy>=2.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
tiktoken>=0.5.0
